Initial repository creation.
This bundle establishes the building of a dynamic shared-object library in
and .so file, and compiles a driver that uses this libeary. The library has
nothing (meaningful) in it at the moment.

The file "intfit.h" has been populated with a class that is meant to act as
a term in the functional form of an approximation.
IN 2021/01/13 "master"