Added a temporary debugging preprocessor directive; this will be turned to _DEBUG2_ once development of the system assembly is finished.
Added the private method "evalTerm()" to the fit class so that a term can
be evaluated. This method is __inline-ed, which may not be part of the C++
standard or valid across compilers.

Modified the system assembly to store the matrix element and the vector
element. Debugging diagnostics were injected to "draw" the system on screen
as it is being assembled. (I have intentionally skipped optimizing the use
of indices so that the legic is presently clear in the code. Also, the
matrix is symmetric, but I am assembling it naively to maintain the visual
representation of the assembly.)
IN 2021/01/14 "master"