Added the point-to-double "tp" to the multi-segment class; this is to store the temporary storage needed to perform row-pivots during solution. The pointer is initialized and memory allocated as needed.
Several diagnostics were added and changed, and some were modified.

A function to perform Gaussian elimination with pivoting was added to solve
the system. This function is not part of any class and it is compiled in the
C++ code. This function was taken from my utilities and was modified to be
used here. The function is called in the "compute()" method to solve the
least-squares system. (The choice of not taking advantage of the system's
symmetry is because this property will be lost when arbitrary constraints
will be added for inter-connecting the segment fits.)
IN 2021/01/15 "master"