Added the notion of a constraint into the least-squares system via a struct construct. The "struct inTFit_constraint_s" is part of the multi-fit class. It contains the type of constraint, which is specified via an "enum-ed int" and can be one of null, value, and derivative. The null type is so that a dummy constraint can be added, in case the process is done programmatically. The value and derivative constraints are for fixing the value of the two fits or the derivatives of the at the specified point via an equation. The "sign" specifies what the sign of the second fit's terms should be; this is to provide with enough generality into the constraint. The right hand side can allow for an "inhomogeneous constraint", which can be something like: "fit 1 at x + sign*(fit2 at x) = rhs" for the given independent variable x.
A std::vector<> of the constraint structs was added as well as an integer
to keep track of the count. (Those are private variables.)

The "clear()" method was augmented to clean up the constraints.

The public method "addConstraint( struct )" was added to the multi-fit
class. This method performs some consistency checks and returns an error.

A place-holder was put in the method that allocates the memory for the LS
system that is to be solved.
IN 2021/01/18 "master"